added linkList config
