+++ src/components/Footer.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { yandexPraktikumLink } from '../utils/linksList'
 
 const Footer = () => {
   return (
@@ -6,8 +7,8 @@ const Footer = () => {
       <p className="footer__copyrights">&#169; Маршак, 2020.</p>
       <p className="footer__made-by">Сделано студентами&nbsp;
         <a className="footer__link"
-           href="https://praktikum.yandex.ru/"
-           rel="noreferrer" target="_blank">Яндекс.Практикум</a></p>
+           href={yandexPraktikumLink.link}
+           rel="noreferrer" target="_blank">{yandexPraktikumLink.title}</a></p>
     </footer>
   )
 }
