+++ src/utils/linksList.js
@@ -1,0 +1,22 @@
+const marshakLogoLink = 'https://marshakbooks.ru/';
+
+const yandexPraktikumLink = {
+  title: 'Яндекс.Практикума',
+  link: 'https://www.notion.so/de0ecc64f8134e178448dcbc6382046e',
+}
+
+const offerLink = '#';
+
+/* TODO После добавления первого трека проекта отредактировать ссылки
+    чтобы они вели на трек */
+const streamingsList = [
+  {link: 'https://music.yandex.ru/', name: 'Яндекс.Музыка', id: 0},
+  {link: 'https://www.spotify.com/ru-ru/', name: 'Spotify', id: 1},
+  {link: 'https://www.apple.com/ru/apple-music/', name: 'Apple Music', id: 2},
+  {link: 'https://vk.com/vkmusic', name: 'VK Music', id: 3},
+];
+
+export { marshakLogoLink, yandexPraktikumLink, streamingsList, offerLink };
+
+
+
