--- src/utils/streamingsLinkList.js
@@ -1,11 +1,0 @@
-// Массив для списка стримингов.
-// Для добавления новой ссылки, создайте новый объект.
-
-const streamingsList = [
-  {link: 'https://music.yandex.ru/', name: 'Яндекс.Музыка', id: 0},
-  {link: 'https://www.spotify.com/ru-ru/', name: 'Spotify', id: 1},
-  {link: 'https://www.apple.com/ru/apple-music/', name: 'Apple Music', id: 2},
-  {link: 'https://vk.com/vkmusic', name: 'VK Music', id: 3},
-];
-
-export { streamingsList };
